Plane: moved is_flying related code to its own file
